+++ apps/cugetreg/src/lib/mapper.ts
@@ -1,5 +1,6 @@
 import type { Day } from '@cugetreg/utils/types';
 import type { GenEdType, SortBy, StudyProgram } from '@cugetreg/zod-schemas';
+import type { Semester } from '@cugetreg/zod-schemas';
 
 export function studyProgramMapper(studyProgram: StudyProgram) {
   switch (studyProgram) {
@@ -60,5 +61,47 @@ export function sortByMapper(sortBy: SortBy) {
       return 'รหัสวิชา';
     default:
       throw new Error(`SortBy ${sortBy} is invalid`);
+  }
+}
+
+export function normalizeDayMapper(d: string | number | undefined | null): Day | undefined {
+  if (d === undefined || d === null) return undefined;
+  
+  const up = String(d).toUpperCase();
+  if (up.startsWith('MO') || up === '1') return 'MO';
+  if (up.startsWith('TU') || up === '2') return 'TU';
+  if (up.startsWith('WE') || up === '3') return 'WE';
+  if (up.startsWith('TH') || up === '4') return 'TH';
+  if (up.startsWith('FR') || up === '5') return 'FR';
+  if (up.startsWith('SA') || up === '6') return 'SA';
+  if (up.startsWith('SU') || up === '7' || up === '0') return 'SU';
+  
+  const KNOWN_DAYS = ['MO', 'TU', 'WE', 'TH', 'FR', 'SA', 'SU'];
+  const sliced = up.slice(0, 2);
+  return KNOWN_DAYS.includes(sliced) ? (sliced as Day) : undefined;
+}
+
+export function normalizeGenedMapper(g: string | undefined | null): string[] {
+  if (!g || g === 'NO') return [];
+  
+  const up = String(g).toUpperCase();
+  if (up.startsWith('SC')) return ['SC'];
+  if (up.startsWith('SO')) return ['SO'];
+  if (up.startsWith('HU')) return ['HU'];
+  if (up.startsWith('IN')) return ['IN'];
+  
+  return [up];
+}
+
+export function semesterToTermMapper(semester: Semester | string): string {
+  switch (semester) {
+    case 'FIRST':
+      return '1';
+    case 'SECOND':
+      return '2';
+    case 'SUMMER':
+      return '3';
+    default:
+      return '1';
   }
 }
